Debt/193 api test coverage - routes - request and users route

+++ src/models/user/user.notifications/user.notifications.methods.ts
@@ -9,7 +9,7 @@ import { IUserDocument } from "../user.types";
  * @return {Promise<INotificationDocument[]>}
  */
 export async function getNotifications(
-  this: IUserDocument,
+  this: IUserDocument
 ): Promise<INotificationDocument[]> {
   return NotificationModel.find({
     "targetId": this.id,
@@ -24,11 +24,22 @@ export async function getNotifications(
  */
 export async function dismissNotification(
   this: IUserDocument,
-  notificationId: string,
+  notificationId: string
 ): Promise<INotificationDocument[]> {
+  const notificationDocument = await NotificationModel.findById(notificationId);
+
+  if (!notificationDocument) {
+    throw new Error("Unable to find notification by id");
+  }
+  if (notificationDocument.targetId !== this.id) {
+    throw new Error(
+      "Dismiss notifications: illegal operation - targetId and userId don't match"
+    );
+  }
+
   await NotificationModel.findByIdAndDelete(notificationId);
-  this.notifications = this.notifications.filter((notification) =>
-    notification !== notificationId,
+  this.notifications = this.notifications.filter(
+    (notification) => notification !== notificationId
   );
   this.markModified("notifications");
   await this.save();
